Refactor API configuration and structure
- **Centralized PostgreSQL Configuration**
  Utilize `restalchemy.common.config_opts` to register PostgreSQL database options, replacing redundant local configurations. This streamlines database setup across services.
  - Updated in `gservice.py`, `orch_api.py`, and `user_api.py`.

- **Restructured API Modules**
  Reorganized API code into `api` subdirectories for better project structure:
  - Moved `app.py` and related modules to `genesis_core/{orch_api,user_api}/api/`.
  - Introduced shared `common/api/controllers.py` for reusable controllers.

- **Refined Engine Configuration**
  Updated database engine initialization to use `configure_postgresql_factory(CONF)`, aligning with centralized configuration practices.

- **Miscellaneous Improvements**
  - Added missing copyright headers.
  - Marked `HealthController` with a TODO for future optimization.
  - Cleaned up CLI option formatting and imports.

- **Simplified Maintenance**: Shared configurations and controllers reduce redundancy.
- **Consistent Structure**: APIs follow a unified module layout.
- **Scalability**: Centralized configuration eases future database adjustments.
