+++ genesis_core/cmd/gservice.py
@@ -13,10 +13,12 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+
 import logging
 import sys
 
 from oslo_config import cfg
+from restalchemy.common import config_opts as ra_config_opts
 from restalchemy.storage.sql import engines
 
 from genesis_core.common import config
@@ -24,19 +26,10 @@ from genesis_core.common import log as infra_log
 from genesis_core.gservice.service import GeneralService
 
 
-db_opts = [
-    cfg.StrOpt(
-        "connection-url",
-        default="postgresql://genesis_core:genesis_core@127.0.0.1:5432/genesis_core",
-        help="Connection URL to db",
-    ),
-]
-
-
 DOMAIN = "gservice"
 
 CONF = cfg.CONF
-CONF.register_opts(db_opts, "db")
+ra_config_opts.register_posgresql_db_opts(CONF)
 
 
 def main():
@@ -47,7 +40,7 @@ def main():
     infra_log.configure()
     log = logging.getLogger(__name__)
 
-    engines.engine_factory.configure_factory(db_url=CONF.db.connection_url)
+    engines.engine_factory.configure_postgresql_factory(CONF)
 
     service = GeneralService(iter_min_period=0.5)
