+++ genesis_core/cmd/user_api.py
@@ -1,39 +1,57 @@
+#    Copyright 2025 Genesis Corporation.
+#
+#    All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
 import logging
 import sys
 
 from gcl_looper.services import bjoern_service
 from oslo_config import cfg
+from restalchemy.common import config_opts as ra_config_opts
 from restalchemy.storage.sql import engines
 
-from genesis_core.user_api import app
+from genesis_core.user_api.api import app
 from genesis_core.common import config
 from genesis_core.common import log as infra_log
 
 
 api_cli_opts = [
     cfg.StrOpt(
-        "bind-host", default="127.0.0.1", help="The host IP to bind to"
+        "bind-host",
+        default="127.0.0.1",
+        help="The host IP to bind to",
     ),
-    cfg.IntOpt("bind-port", default=11010, help="The port to bind to"),
     cfg.IntOpt(
-        "workers", default=1, help="How many http servers should be started"
+        "bind-port",
+        default=11010,
+        help="The port to bind to",
     ),
-]
-
-db_opts = [
-    cfg.StrOpt(
-        "connection-url",
-        default="postgresql://genesis_core:genesis_core@127.0.0.1:5432/genesis_core",
-        help="Connection URL to db",
+    cfg.IntOpt(
+        "workers",
+        default=1,
+        help="How many http servers should be started",
     ),
 ]
 
 
 DOMAIN = "user_api"
 
+
 CONF = cfg.CONF
 CONF.register_cli_opts(api_cli_opts, DOMAIN)
-CONF.register_opts(db_opts, "db")
+ra_config_opts.register_posgresql_db_opts(CONF)
 
 
 def main():
@@ -44,7 +62,7 @@ def main():
     infra_log.configure()
     log = logging.getLogger(__name__)
 
-    engines.engine_factory.configure_factory(db_url=CONF.db.connection_url)
+    engines.engine_factory.configure_postgresql_factory(CONF)
 
     log.info(
         "Start service on %s:%s",
