+++ genesis_core/user_api/api/app.py
@@ -22,33 +22,31 @@ from restalchemy.api.middlewares import logging as logging_mw
 from restalchemy.openapi import structures as openapi_structures
 from restalchemy.openapi import engines as openapi_engines
 
-from genesis_core.orch_api import controllers
-from genesis_core.orch_api import routes as app_routes
-from genesis_core.orch_api import versions
+from genesis_core.user_api.api import routes as app_routes
+from genesis_core.user_api.api import versions
 from genesis_core import version
 
 
-class OrchApiApp(routes.Route):
-    __controller__ = controllers.RootController
-    __allow_methods__ = [routes.FILTER]
+class UserApiApp(routes.RootRoute):
+    pass
 
 
 # Route to /v1/ endpoint.
 setattr(
-    OrchApiApp,
+    UserApiApp,
     versions.API_VERSION_v1,
     routes.route(app_routes.ApiEndpointRoute),
 )
 
 
 def get_api_application():
-    return OrchApiApp
+    return UserApiApp
 
 
 def get_openapi_engine():
     openapi_engine = openapi_engines.OpenApiEngine(
         info=openapi_structures.OpenApiInfo(
-            title=f"Genesis Core {versions.API_VERSION_v1} Orch API",
+            title=f"Genesis Core {versions.API_VERSION_v1} User API",
             version=version.version_info.release_string(),
             description=f"OpenAPI - Genesis Core {versions.API_VERSION_v1}",
         ),
